[Refactored] Added 'pages' folder && index.js on 'Components'
