+++ src/components/index.js
@@ -1,0 +1,4 @@
+export { default as SearchBar } from './SearchBar';
+export { default as ProductsList } from './ProductsList';
+export { default as CategoriesList } from './CategoriesList';
+export { default as ShoppingCartButton } from './ShoppingCartButton';
